+++ CHANGELOG.md
@@ -30,7 +30,13 @@ These changes are available on the `master` branch, but have not yet been releas
   ([#2587](https://github.com/Pycord-Development/pycord/pull/2587/))
 - Added optional `filter` parameter to `utils.basic_autocomplete()`.
   ([#2590](https://github.com/Pycord-Development/pycord/pull/2590))
-- Added soundboard support :tada:
+- Added missing `with_counts` parameter to `fetch_guilds` method.
+  ([#2615](https://github.com/Pycord-Development/pycord/pull/2615))
+- Added missing permissions: `Permissions.use_soundboard`,
+  `Permissions.use_external_sounds` and
+  `Permissions.view_creator_monetization_analytics`.
+  ([#2620](https://github.com/Pycord-Development/pycord/pull/2620))
+- Added soundboard features
   - Manage guild soundboard sounds with `Guild.fetch_sounds()`, `Guild.create_sound()`,
     `SoundboardSound.edit()`, and `SoundboardSound.delete()`
   - Access Discord's default sounds with `Client.fetch_default_sounds()`
@@ -38,7 +44,7 @@ These changes are available on the `master` branch, but have not yet been releas
   - New `on_voice_channel_effect_send` event for sound and emoji effects
   - Soundboard limits based on guild premium tier (8-48 slots) in
     `Guild.soundboard_limit`
-    ([#2623](https://github.com/Pycord-Development/pycord/pull/2623))
+  ([#2623](https://github.com/Pycord-Development/pycord/pull/2623))
 
 ### Fixed